+++ libft/ft_utoa_base.c
@@ -6,17 +6,17 @@
 /*   By: lkaser <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/21 14:38:16 by lkaser            #+#    #+#             */
-/*   Updated: 2017/12/29 12:45:43 by lkaser           ###   ########.fr       */
+/*   Updated: 2017/12/29 17:27:01 by lkaser           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_utoa_base(unsigned long long nbr, char *base_str, unsigned base)
+char	*ft_utoa_base(uintmax_t nbr, char *base_str, unsigned base)
 {
-	unsigned long long	n;
-	unsigned			digits;
-	char				*str;
+	uintmax_t 	n;
+	unsigned	digits;
+	char		*str;
 
 	digits = (nbr == 0);
 	n = nbr;
